Move functions only used in resample.py to resample.py.
