+++ harmony_regridding_service/resample.py
@@ -1,20 +1,18 @@
 """Module for resampling functions."""
 
 from logging import getLogger
+from pathlib import PurePath
 
 import numpy as np
 from netCDF4 import (
     Dataset,
+    Dimension,
 )
 from pyresample.ewa import DaskEWAResampler
 from pyresample.geometry import AreaDefinition
 from varinfo import VarInfoFromNetCDF4
 
 from harmony_regridding_service.dimensions import (
-    copy_1d_dimension_variables,
-    copy_dimensions,
-    create_dimension,
-    get_all_dimensions,
     get_column_dims,
     get_row_dims,
     horizontal_dims_for_variable,
@@ -28,7 +26,7 @@ from harmony_regridding_service.grid import compute_source_swath
 from harmony_regridding_service.utilities import (
     copy_var_with_attrs,
     copy_var_without_metadata,
-    integer_like,
+    get_bounds_var,
 )
 
 logger = getLogger(__name__)
@@ -143,6 +141,11 @@ def resampler_kwargs(data: np.ndarray, original_dtype: np.dtype) -> dict:
         kwargs['maximum_weight_mode'] = True
 
     return kwargs
+
+
+def integer_like(test_type: np.dtype) -> bool:
+    """Return True if the datatype is integer like."""
+    return np.issubdtype(np.dtype(test_type), np.integer)
 
 
 def copy_resampled_bounds_variable(
@@ -395,3 +398,117 @@ def needs_rotation(var_info: VarInfoFromNetCDF4, variable: str) -> bool:
         needs_rotation = True
 
     return needs_rotation
+
+
+def copy_dimensions(
+    dimensions: set[str], source_ds: Dataset, target_ds: Dataset
+) -> set[str]:
+    """Copy each dimension from source to target.
+
+    ensure the first dimensions copied are the UNLIMITED dimensions.
+    """
+
+    def sort_unlimited_first(dimension_name):
+        """Sort dimensions so that unlimited are first in list."""
+        the_dim = get_dimension(source_ds, dimension_name)
+        return not the_dim.isunlimited()
+
+    sorted_dims = sorted(list(dimensions), key=sort_unlimited_first)
+
+    for dim in sorted_dims:
+        copy_dimension(dim, source_ds, target_ds)
+
+
+def copy_dimension(dimension_name: str, source_ds: Dataset, target_ds: Dataset) -> str:
+    """Copy dimension from source to target file."""
+    source_dimension = get_dimension(source_ds, dimension_name)
+
+    source_size = None
+    if not source_dimension.isunlimited():
+        source_size = source_dimension.size
+
+    dim = PurePath(dimension_name)
+    target_group = target_ds.createGroup(dim.parent)
+    return target_group.createDimension(dim.name, source_size)
+
+
+def get_dimension(dataset: Dataset, dimension_name: str) -> Dimension:
+    """Return a dimension object for a dimension name.
+
+    Return the Dimension for an arbitrarily nested dimension name.
+    """
+    dim = PurePath(dimension_name)
+    return dataset.createGroup(dim.parent).dimensions[dim.name]
+
+
+def create_dimension(dataset: Dataset, dimension_name: str, size: int) -> Dimension:
+    """Create a fully qualified dimension on the dataset."""
+    dim = PurePath(dimension_name)
+    group = dataset.createGroup(dim.parent)
+    return group.createDimension(dim.name, size)
+
+
+def get_all_dimensions(var_info: VarInfoFromNetCDF4) -> set[str]:
+    """Return a list of all dimensions in the file."""
+    dimensions = set()
+    for variable_name in var_info.get_all_variables():
+        variable = var_info.get_variable(variable_name)
+        for dim in variable.dimensions:
+            dimensions.add(dim)
+
+    return dimensions
+
+
+def copy_1d_dimension_variables(
+    source_ds: Dataset,
+    target_ds: Dataset,
+    dim_var_names: set[str],
+    target_area: AreaDefinition,
+    var_info: VarInfoFromNetCDF4,
+) -> set[str]:
+    """Copy 1 dimensional dimension variables.
+
+    These are the variables associated directly with the resampled
+    longitudes, latitudes, Columns, rows, x-variables, and y-variables.
+    """
+    # pylint: disable-msg=too-many-locals
+    one_d_vars = {
+        dim_var_name
+        for dim_var_name in dim_var_names
+        if len(var_info.get_variable(dim_var_name).dimensions) == 1
+    }
+
+    xdims = get_column_dims(one_d_vars, var_info)
+    ydims = get_row_dims(one_d_vars, var_info)
+
+    for dim_name in one_d_vars:
+        if dim_name in xdims:
+            target_coords = target_area.projection_x_coords
+            standard_metadata = {
+                'long_name': 'longitude',
+                'standard_name': 'longitude',
+                'units': 'degrees_east',
+            }
+        elif dim_name in ydims:
+            target_coords = target_area.projection_y_coords
+            standard_metadata = {
+                'long_name': 'latitude',
+                'standard_name': 'latitude',
+                'units': 'degrees_north',
+            }
+        else:
+            raise RegridderException(
+                f'dim_name: {dim_name} not found in projection dimensions'
+            )
+
+        (_, t_var) = copy_var_without_metadata(source_ds, target_ds, dim_name)
+
+        bounds_var = get_bounds_var(var_info, dim_name)
+
+        if bounds_var:
+            standard_metadata['bounds'] = bounds_var
+        t_var.setncatts(standard_metadata)
+
+        t_var[:] = target_coords
+
+    return one_d_vars
